feat: add MongoDB cache validation to show endpoint

+++ src/main/java/com/tvnova/client/TvMazeClient.java
@@ -33,7 +33,7 @@ public class TvMazeClient {
     }
   }
 
-  public TvMazeShow getShowById(Integer id) {
+  public TvMazeShow getShowById(long id) {
     try {
       return tvMazeRestClient.get()
         .uri("/shows/{id}", id)
