+++ src/main/java/com/tvnova/service/ITvMazeService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface ITvMazeService {
   List<ShowResponse> searchShows(String query);
 
-  ShowResponse getShowById(Integer id);
+  ShowResponse getShowById(long id);
 }
